refactor: extract Value type from read.go into value.go
Move ValueKind type and constants, Value struct and all its methods
(IsNull, Kind, String, Bool, Int64, Float64, RawString, Text,
TextFromUTF16, Name, Key, Keys, Index, Len, Reader), and
errorReadCloser struct into a new value.go (264 lines).

read.go shrinks from 981 to 728 lines. objfmt stays in read.go
as package-wide infrastructure used by XRef parser, encrypt.go,
and Value.String(). value.go references objfmt, applyFilter, and
resolve across the same package without any import changes.
