+++ read.go
@@ -520,65 +520,6 @@ func findStartxrefOffset(f io.ReaderAt, size int64) (int64, error) {
 	return pos, nil
 }
 
-// A Value is a single PDF value, such as an integer, dictionary, or array.
-// The zero Value is a PDF null (Kind() == Null, IsNull() = true).
-type Value struct {
-	r    *Reader
-	ptr  objptr
-	data interface{}
-}
-
-// IsNull reports whether the value is a null. It is equivalent to Kind() == Null.
-func (v Value) IsNull() bool {
-	return v.data == nil
-}
-
-// A ValueKind specifies the kind of data underlying a Value.
-type ValueKind int
-
-// The PDF value kinds.
-const (
-	Null ValueKind = iota
-	Bool
-	Integer
-	Real
-	String
-	Name
-	Dict
-	Array
-	Stream
-)
-
-// Kind reports the kind of value underlying v.
-func (v Value) Kind() ValueKind {
-	switch v.data.(type) {
-	default:
-		return Null
-	case bool:
-		return Bool
-	case int64:
-		return Integer
-	case float64:
-		return Real
-	case string:
-		return String
-	case name:
-		return Name
-	case dict:
-		return Dict
-	case array:
-		return Array
-	case stream:
-		return Stream
-	}
-}
-
-// String returns a textual representation of the value v.
-// Note that String is not the accessor for values with Kind() == String.
-// To access such values, see RawString, Text, and TextFromUTF16.
-func (v Value) String() string {
-	return objfmt(v.data)
-}
 
 func objfmt(x interface{}) string {
 	switch x := x.(type) {
@@ -638,154 +579,6 @@ func objfmt(x interface{}) string {
 	}
 }
 
-// Bool returns v's boolean value.
-// If v.Kind() != Bool, Bool returns false.
-func (v Value) Bool() bool {
-	x, ok := v.data.(bool)
-	if !ok {
-		return false
-	}
-	return x
-}
-
-// Int64 returns v's int64 value.
-// If v.Kind() != Int64, Int64 returns 0.
-func (v Value) Int64() int64 {
-	x, ok := v.data.(int64)
-	if !ok {
-		return 0
-	}
-	return x
-}
-
-// Float64 returns v's float64 value, converting from integer if necessary.
-// If v.Kind() != Float64 and v.Kind() != Int64, Float64 returns 0.
-func (v Value) Float64() float64 {
-	x, ok := v.data.(float64)
-	if !ok {
-		x, ok := v.data.(int64)
-		if ok {
-			return float64(x)
-		}
-		return 0
-	}
-	return x
-}
-
-// RawString returns v's string value.
-// If v.Kind() != String, RawString returns the empty string.
-func (v Value) RawString() string {
-	x, ok := v.data.(string)
-	if !ok {
-		return ""
-	}
-	return x
-}
-
-// Text returns v's string value interpreted as a “text string” (defined in the PDF spec)
-// and converted to UTF-8.
-// If v.Kind() != String, Text returns the empty string.
-func (v Value) Text() string {
-	x, ok := v.data.(string)
-	if !ok {
-		return ""
-	}
-	if isPDFDocEncoded(x) {
-		return pdfDocDecode(x)
-	}
-	if isUTF16(x) {
-		return utf16Decode(x[2:])
-	}
-	return x
-}
-
-// TextFromUTF16 returns v's string value interpreted as big-endian UTF-16
-// and then converted to UTF-8.
-// If v.Kind() != String or if the data is not valid UTF-16, TextFromUTF16 returns
-// the empty string.
-func (v Value) TextFromUTF16() string {
-	x, ok := v.data.(string)
-	if !ok {
-		return ""
-	}
-	if len(x)%2 == 1 {
-		return ""
-	}
-	if x == "" {
-		return ""
-	}
-	return utf16Decode(x)
-}
-
-// Name returns v's name value.
-// If v.Kind() != Name, Name returns the empty string.
-// The returned name does not include the leading slash:
-// if v corresponds to the name written using the syntax /Helvetica,
-// Name() == "Helvetica".
-func (v Value) Name() string {
-	x, ok := v.data.(name)
-	if !ok {
-		return ""
-	}
-	return string(x)
-}
-
-// Key returns the value associated with the given name key in the dictionary v.
-// Like the result of the Name method, the key should not include a leading slash.
-// If v is a stream, Key applies to the stream's header dictionary.
-// If v.Kind() != Dict and v.Kind() != Stream, Key returns a null Value.
-func (v Value) Key(key string) Value {
-	x, ok := v.data.(dict)
-	if !ok {
-		strm, ok := v.data.(stream)
-		if !ok {
-			return Value{}
-		}
-		x = strm.hdr
-	}
-	return v.r.resolve(v.ptr, x[name(key)])
-}
-
-// Keys returns a sorted list of the keys in the dictionary v.
-// If v is a stream, Keys applies to the stream's header dictionary.
-// If v.Kind() != Dict and v.Kind() != Stream, Keys returns nil.
-func (v Value) Keys() []string {
-	x, ok := v.data.(dict)
-	if !ok {
-		strm, ok := v.data.(stream)
-		if !ok {
-			return nil
-		}
-		x = strm.hdr
-	}
-	keys := []string{} // not nil
-	for k := range x {
-		keys = append(keys, string(k))
-	}
-	sort.Strings(keys)
-	return keys
-}
-
-// Index returns the i'th element in the array v.
-// If v.Kind() != Array or if i is outside the array bounds,
-// Index returns a null Value.
-func (v Value) Index(i int) Value {
-	x, ok := v.data.(array)
-	if !ok || i < 0 || i >= len(x) {
-		return Value{}
-	}
-	return v.r.resolve(v.ptr, x[i])
-}
-
-// Len returns the length of the array v.
-// If v.Kind() != Array, Len returns 0.
-func (v Value) Len() int {
-	x, ok := v.data.(array)
-	if !ok {
-		return 0
-	}
-	return len(x)
-}
 
 // resolveInStream locates ptr inside an ObjStm (object stream) by scanning
 // the index entries, following Extends chains as needed.
@@ -861,52 +654,6 @@ func (r *Reader) resolve(parent objptr, x interface{}) Value {
 	}
 }
 
-type errorReadCloser struct {
-	err error
-}
-
-func (e *errorReadCloser) Read([]byte) (int, error) {
-	return 0, e.err
-}
-
-func (e *errorReadCloser) Close() error {
-	return e.err
-}
-
-// Reader returns the data contained in the stream v.
-// If v.Kind() != Stream, Reader returns a ReadCloser that
-// responds to all reads with a “stream not present” error.
-func (v Value) Reader() io.ReadCloser {
-	x, ok := v.data.(stream)
-	if !ok {
-		return &errorReadCloser{fmt.Errorf("stream not present")}
-	}
-	streamLen := v.Key("Length").Int64()
-	if streamLen == 0 {
-		return io.NopCloser(bytes.NewReader(nil))
-	}
-	var rd io.Reader
-	rd = io.NewSectionReader(v.r.f, x.offset, streamLen)
-	if v.r.key != nil {
-		rd = decryptStream(v.r.key, v.r.useAES, x.ptr, rd)
-	}
-	filter := v.Key("Filter")
-	param := v.Key("DecodeParms")
-	switch filter.Kind() {
-	default:
-		panic(fmt.Errorf("unsupported filter %v", filter))
-	case Null:
-		// ok
-	case Name:
-		rd = applyFilter(rd, filter.Name(), param)
-	case Array:
-		for i := 0; i < filter.Len(); i++ {
-			rd = applyFilter(rd, filter.Index(i).Name(), param.Index(i))
-		}
-	}
-
-	return io.NopCloser(rd)
-}
 
 func applyFilter(rd io.Reader, name string, param Value) io.Reader {
 	switch name {
